Add Jest tests and update scripts
