+++ README.md
@@ -47,6 +47,12 @@ cp .env.example .env
 npm start
 ```
 
+### 5. Запустить тесты
+
+```bash
+npm test
+```
+
 Перед запуском убедитесь, что в `.env` указан `DATABASE_URL` с данными PostgreSQL
 и при необходимости `PGSSL`. Таблицы будут созданы автоматически при первом запу
 ске.
