+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "start": "node app.js"
   },
   "keywords": [],
@@ -11,13 +11,16 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "body-parser": "^2.2.0",
+    "connect-pg-simple": "^8.0.0",
     "ejs": "^3.1.10",
     "express": "^4.21.2",
-    "method-override": "^3.0.0",
-    "pg": "^8.11.1",
     "express-session": "^1.17.3",
-    "connect-pg-simple": "^8.0.0",
-    "bcryptjs": "^2.4.3"
+    "method-override": "^3.0.0",
+    "pg": "^8.11.1"
+  },
+  "devDependencies": {
+    "jest": "^30.0.2"
   }
 }
